pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wireguard-vpn-manager"
version = "1.0.0"
description = "WireGuard VPN Orchestration Tool"
authors = [{name = "Grigory Streltsov", email = "strelga@gmail.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PyYAML>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["vpn_manager"]

[project.optional-dependencies]
dev = [
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-PyYAML",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-snapshot>=0.9.0",
]

[project.scripts]
vpn-manager = "vpn_manager.manager:main"

[project.urls]
Homepage = "https://github.com/strelga/wireguard-vpn-manager"
Repository = "https://github.com/strelga/wireguard-vpn-manager.git"
Issues = "https://github.com/strelga/wireguard-vpn-manager/issues"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.8+
target-version = "py313"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = [
    "E4",   # Import errors
    "E7",   # Statement errors
    "E9",   # Runtime errors
    "F",    # Pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "T20",  # flake8-print
    "PL",   # Pylint
]
ignore = [
    "E501",  # Line too long (handled by formatter)
    "T201",  # Allow print statements
    "T203",  # Allow pprint
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Per-file ignores
[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["PLR0913", "PLR2004"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["vpn-manager"]
